refactor: 컴포넌트, 페이지 파일 구분
