+++ src/mocks/faqData.json
@@ -8,11 +8,11 @@
     "answer": "저희 ‘아차’는 등록된 대학의 경우 이용이 가능합니다! 앱 설치 후 대학이 포함되어 있는지 찾아보세요저희 ‘아차’는 등록된 대학의 경우 이용이 가능합니다! 앱 설치 후 대학이 포함되어 있는지 찾아보세요저희 ‘아차’는 등록된 대학의 경우 이용이 가능합니다! 앱 설치 후 대학이 포함되어 있는지 찾아보세요저희 ‘아차’는 등록된 대학의 경우 이용이 가능합니다! 앱 설치 후 대학이 포함되어 있는지 찾아보세요저희 ‘아차’는 등록된 대학의 경우 이용이 가능합니다! 앱 설치 후 대학이 포함되어 있는지 찾아보세요저희 ‘아차’는 등록된 대학의 경우 이용이 가능합니다! 앱 설치 후 대학이 포함되어 있는지 찾아보세요"
   },
   {
-    "question": "질문3",
-    "answer": "답변3"
+    "question": "어떻게 가입하나요?",
+    "answer": "아차 회원여부에 따라 로그인 시에 회원가입이 이루어지고 있습니다. 로그인을 통해서 회원가입을 진행해보세요!"
   },
   {
-    "question": "질문4",
-    "answer": "답변4"
+    "question": "알림을 어떻게 추가/관리하나요?",
+    "answer": "알림을 추가하는 기능은 추후에 개발될 예정입니다. 현재 아차 '앱'을 통해서 알림을 끄고 킬 수 있습니다."
   }
 ]
